requirements.txt
amqp==5.1.1
asgiref==3.7.2
async-timeout==4.0.2
billiard==3.6.4.0
celery==5.2.7
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
Django==4.2.1
django-redis==5.2.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
kombu==5.3.0
# numpy==1.24.3
pip-autoremove==0.10.0
prompt-toolkit==3.0.38
psycopg2-binary
# psycopg2==2.9.6
PyJWT==2.7.0
pytz==2023.3
redis==4.5.5
rel==0.4.9
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.6.3
tzdata
vine==5.0.0
wcwidth==0.2.6
websocket-client==1.5.2